Move metadata plugin registration to seperate package
